requirements.txt
flask==2.0.1
moviepy==1.0.3
Pillow==9.0.0
python-dotenv==0.19.0
werkzeug==2.0.1 